requirements.txt
# ============================================================
# requirements.txt — chatbot HaUI
# Tested: Python 3.11, Anaconda conda 26.1.1, Windows
#
# Cài đặt:
#   conda create -n haui-bot python=3.11 -y
#   conda activate haui-bot
#   pip install -r requirements.txt
# ============================================================

# ── LLM ──────────────────────────────────────────────────────
anthropic==0.84.0

# ── Vector store ─────────────────────────────────────────────
# chromadb 1.5.x dùng abi3 wheel, hoạt động tốt trên Python 3.11
chromadb==1.5.5

# ── Embedding model ───────────────────────────────────────────
# sentence-transformers 3.x yêu cầu Python >= 3.10
# KHÔNG dùng bản 4.x+ vì đổi API, có thể break code
sentence-transformers==3.4.1

# ── PyTorch CPU (bắt buộc cho sentence-transformers) ─────────
# Nếu máy có GPU NVIDIA thì thay bằng lệnh riêng:
#   pip install torch --index-url https://download.pytorch.org/whl/cu121
torch==2.5.1
torchvision==0.20.1

# ── LangChain text splitter ───────────────────────────────────
# Chỉ cần module này, không cần cài toàn bộ LangChain
langchain-text-splitters==0.3.8

# ── Markdown frontmatter parser ───────────────────────────────
python-frontmatter==1.1.0

# ── Biến môi trường (.env file) ───────────────────────────────
python-dotenv==1.0.1

# ── Tiện ích ──────────────────────────────────────────────────
tqdm==4.67.1          # progress bar khi embed
numpy==1.26.4         # QUAN TRỌNG: chromadb + sentence-transformers cần numpy < 2.0
rank-bm25>=0.2.2          # BM25 search
streamlit>=1.32.0
FlagEmbedding>=1.3.0         # bge-reranker-v2-m3